6_Agents/app/requirements.txt
openai
chromadb
streamlit
python-dotenv
